refactor(agent): extract adr-010 prompt-fencing primitives into prompt_fence
PR-5 step 1 (drop agentic surfaces): pulls fenced/JOB_CAP/RESUME_CAP/
FENCE_TAG_PATTERNS/neutralize_transcript_boundaries out of agent::tools into a
new dependency-free crate::prompt_fence module so the pipeline, extension
bridge, and structured-output re-ask path keep their injection defense once
agent/ is deleted in step 2. AUTOPILOT_NOTE_SYSTEM moves into
autopilot_helpers, its only remaining consumer.

Pure move: every existing test travels with its primitive, unchanged, plus a
mutation check (verified, then reverted) confirming a broken fenced() still
fails pipeline::resume::test::every_untrusted_block_is_fenced_and_forgery_resistant.
ECHO_CAP/clamped_echo stay in agent::tools — no surviving consumer needs them.

tests/architecture.rs: classifies prompt_fence as L0, drops the now-dead
pipeline->agent and autopilot_helpers->agent R7 exceptions, and allowlists
prompt_fence->ipc_contracts for the generated RESUME_CAP constant (same shape
as the existing pipeline/ai_config/scraping exceptions).
